+++ weblate/checks/base.py
@@ -109,13 +109,6 @@ class Check(object):
             or (src not in chars and tgt in chars)
         )
 
-    def check_ends(self, target, ends):
-        """Check whether target ends with one of given ends."""
-        for end in ends:
-            if target.endswith(end):
-                return True
-        return False
-
     def is_language(self, unit, vals):
         """Detect whether language is in given list, ignores variants."""
         return unit.translation.language.base_code in vals