+++ weblate/settings_example.py
@@ -688,6 +688,7 @@ CRISPY_TEMPLATE_PACK = 'bootstrap3'
 #     'weblate.checks.chars.EndSemicolonCheck',
 #     'weblate.checks.chars.MaxLengthCheck',
 #     'weblate.checks.chars.KashidaCheck',
+#     'weblate.checks.chars.PuctuationSpacingCheck',
 #     'weblate.checks.format.PythonFormatCheck',
 #     'weblate.checks.format.PythonBraceFormatCheck',
 #     'weblate.checks.format.PHPFormatCheck',